add eos-firewall-localonly service
This implements a simple iptables/ip6tables firewall which allows the
host to communicate only with resources on the local network, plus a few
explicitly-enabled hosts (the Endless metrics / phone-home servers) and
services (DNS and NTP).

There is a script eos-firewall-localonly which sets up the initial
tables with common local network and local multicast addresses,
permitted interfaces (lo) and the permitted services, and a
NetworkManager dispatch script called eos-firewall-localonly-nm which
adds additional rules allowing traffic to the permitted hosts (so that the
DNS requests are not done at boot time before the network is available)
and any other local network routes which have been assigned via DHCP.

Denied TCP connections are reset so that the local application receives
an error immediately, and any blocked packets are logged to the journal
with a relatively low frequency.

For obvious reasons, the systemd service is disabled by default: this
functionality must be explicitly enabled on the hosts where it is
desirable.